+++ source/main.cpp
@@ -33,7 +33,7 @@ int main()
       /*rand_init=*/true);
 
   // Create SGD optimizer
-  nn::SGD optimizer(&model, /*learning_rate=*/0.001);
+  nn::SGD optimizer(&model, /*learning_rate=*/0.001f);
 
   // Create loss function
   nn::CrossEntropyLoss loss_fn(&model);
@@ -43,10 +43,10 @@ int main()
     // start time counting
     auto start = high_resolution_clock::now();
 
-    double train_correct {};
-    double test_correct {};
-    double train_loss {};
-    double test_loss {};
+    float train_correct {};
+    float test_correct {};
+    float train_loss {};
+    float test_loss {};
 
     std::cout << "------------------ Training -----------------\n";
 
@@ -56,13 +56,13 @@ int main()
       auto logits = model.forward(img->data);
 
       // calculate loss
-      double loss = loss_fn(logits, img->label);
+      float loss = loss_fn(logits, img->label);
       train_loss += loss;
 
       // calculate accuracy
-      size_t pred_label = loss_fn.get_pred().arg_max();
-      size_t true_label = img->label.arg_max();
-      train_correct += (pred_label == true_label) ? 1.0 : 0.0;
+      int pred_label = loss_fn.get_pred().arg_max();
+      int true_label = img->label.arg_max();
+      train_correct += (pred_label == true_label) ? 1.0f : 0.0f;
 
       // backward pass
       loss_fn.backward(img->label);
@@ -81,10 +81,10 @@ int main()
 
     std::cout << "Epoch: " << epoch << "\n"
               << "Train accuracy: "
-              << train_correct / static_cast<double>(train_data.dataset.size())
+              << train_correct / static_cast<float>(train_data.dataset.size())
               << "\n"
               << "Average train loss: "
-              << train_loss / static_cast<double>(train_data.dataset.size())
+              << train_loss / static_cast<float>(train_data.dataset.size())
               << "\n";
 
     // std::cout << "------------------ Testing -----------------\n";
@@ -93,25 +93,25 @@ int main()
       auto logits = model.forward(img->data);
 
       // calculate loss
-      double loss = loss_fn(logits, img->label);
+      float loss = loss_fn(logits, img->label);
       test_loss += loss;
 
       // calculate accuracy
-      size_t pred_label = loss_fn.get_pred().arg_max();
-      size_t true_label = img->label.arg_max();
-      test_correct += (pred_label == true_label) ? 1.0 : 0.0;
+      int pred_label = loss_fn.get_pred().arg_max();
+      int true_label = img->label.arg_max();
+      test_correct += (pred_label == true_label) ? 1.0f : 0.0f;
     }
 
     std::cout << "Test accuracy: "
-              << test_correct / static_cast<double>(test_data.dataset.size())
+              << test_correct / static_cast<float>(test_data.dataset.size())
               << "\n"
               << "Average test loss: "
-              << test_loss / static_cast<double>(test_data.dataset.size())
+              << test_loss / static_cast<float>(test_data.dataset.size())
               << "\n";
 
     // end time counting
     auto end = high_resolution_clock::now();
-    duration<double, std::milli> time = end - start;
+    duration<float, std::milli> time = end - start;
 
     std::cout << "Time: " << time.count() / 60000.0 << " minutes";
   }
